requirements.txt
inference-sdk
ultralytics
opencv-python
RPi.GPIO==0.7.1
rpi_ws281x